FPLA-720 Address updates

+++ service/src/main/java/uk/gov/hmcts/reform/fpl/model/robotics/RoboticsData.java
@@ -22,5 +22,5 @@ public class RoboticsData {
     private final String allocation;
     private final String issueDate;
     private final Applicant applicant;
-    private final long owningCourt;
+    private final int owningCourt;
 }
